Combined string_functions.h and string_utils.h
