+++ src/dagmc.cpp
@@ -1,7 +1,7 @@
 
 #include "openmc/dagmc.h"
 #include "openmc/error.h"
-#include "openmc/string_functions.h"
+#include "openmc/string_utils.h"
 #include "openmc/settings.h"
 #include "openmc/geometry.h"
